move genotype_svs and variant_recall out of Genotyper class
